Cleanign up EditZing
