+++ frontend/src/modules/EditZing/Types/Student.ts
@@ -1,5 +1,3 @@
-import React from 'react'
-
 export type Student = {
   college: string
   courseId: string
